geobuf index

+++ src/geobuf/geobuf.hpp
@@ -196,8 +196,8 @@ struct Decoder
     int precision() const { return std::log10(e); }
     int __dim() const { return dim; }
     std::vector<std::string> __keys() const { return keys; }
-    int __header_size() const { return header_size; }
-    std::vector<int> __offsets() const { return offsets; }
+    uint32_t __header_size() const { return header_size; }
+    std::vector<uint64_t> __offsets() const { return offsets; }
 
   private:
     mapbox::geojson::feature_collection readFeatureCollection(Pbf &pbf);
@@ -212,8 +212,8 @@ struct Decoder
     std::vector<std::string> keys;
 
     const char *head = nullptr;
-    int header_size = -1;
-    std::vector<int> offsets;
+    uint32_t header_size = -1;
+    std::vector<uint64_t> offsets;
 };
 
 } // namespace geobuf
